Added Homework.scala
Homework.scala is the class that should be used for individual homework assignments. This class contains private variables, getters, and setters for the name of the assignment, the due date of the assignment, the time the assignment was completed, the date the assignment was completed, the grade of the completed assignment and a boolean for if the assignment has been completed or not.